Correction types map reduce - Commit qui sert a rien

+++ MapReduce/index/src/IndexReducer.java
@@ -2,6 +2,7 @@ import java.io.IOException;
 import java.util.Iterator;
 
 import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.io.WritableComparable;
 import org.apache.hadoop.mapred.OutputCollector;
@@ -9,9 +10,9 @@ import org.apache.hadoop.mapred.Reporter;
 import org.apache.hadoop.mapreduce.Reducer;
 
 
-public class IndexReducer extends Reducer<Text,IntWritable,Text,Text> {
+public class IndexReducer extends Reducer<Text,LongWritable,Text,Text> {
 
-	public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException,InterruptedException {
+	public void reduce(Text key, Iterable<LongWritable> values, Context context) throws IOException,InterruptedException {
 		// replace KeyType with the real type of your key
 		KeyType key = (KeyType) _key;
